+++ src/datagen.h
@@ -1,0 +1,11 @@
+#ifndef DATAGEN_H
+#define DATAGEN_H 1
+
+#include "types.h"
+
+struct Datagen
+{
+    static void run();
+};
+
+#endif // !DATAGEN_H