+++ src/timer.h
@@ -8,14 +8,16 @@ class Timer
 {
     bool infinite;
     std::chrono::high_resolution_clock::time_point start_time;
-    Time allocated_time;
+    Time allocated_time_min;
+    Time allocated_time_max;
 
 public:
     bool stopped;
 
     void init(const bool infinite, const Time base, const Time increment);
     Time elapsed() const;
-    bool should_stop();
+    bool should_stop_min();
+    bool should_stop_max();
 };
 
 #endif // !TIMER_H